Add support for charging and network constraints for worker

+++ app/src/androidTest/java/xyz/hisname/fireflyiii/fireflydao/DaoFactory.kt
@@ -22,7 +22,7 @@ class DaoFactory {
                     "", randomLong(),
                     AccountAttributes(
                             "", "", "", randomBoolean(), randomAccountType(), "" ,
-                            randomInt(), "", 0.0, "", "",
+                            randomLong(), "", 0.0, "", "",
                             "", "", "", "", "",
                             "", 0.0, "" ,"", "", "",
                             "","", "", randomBoolean()
